Move generated code from src to generated-src
The problem here is that SonarQube hates the generated code and wants all sorts of things corrected as well as line duplication. To avoid this, we will just copy these files to the generated-sources folder, as if they were generated during the build process.

Also changed the responses from JSON/text to XML conforming to the GOBI API format. This will alleviate the edge or other GOBI API consumers from doing JSON/XML conversion.