Updates towards MODGOBI-24
Added a helper method to call mod-configuration to get configured mappings from GOBI orders to FOLIO orders. Added demo schemas for the mappings as well as a helper method to build the map. These will need to be updated/rewritten when the schemas are finalized. Same with the simple unit test.