Added default mappings for ListedElectronicMonograph - MODGOBI-11
In order to allow fields to be mapped differently per order type (electronic/printed/listed-unlisted/etc), we needed to update the mappings schema to allow for mappings per type. In addition, some mappings required additional lookups to map codes to IDs, while other mappings pass codes that will have to be looked up by mod-orders. This leads to some inconsistency that will need to be sorted out. Additionally, translations are in different places, which will need to be addressed in the helper code (which currently uses recursion) in order to execute.

Missing/problematic mappings:
- SOURCE_ID: this mapping doesn't make sense as currently defined.
- VENDOR_REF_NO_TYPE: this mapping is conditional and the current mapping doesn't make sense. It is only set if "vendor_ref_no" is set. We don't have a good way to express conditionals with the current mapping structure.
- "claim_grace": this is retrieved from the vendor interface. We already call this interface to get the vendor ID. We should come up with a mechanism to make retrieved objects available for all calls so that we don't need to make multiple calls to the same API when accessing different fields.
- "expected_receipt_date": same reason as "claim_grace". Also, with electronic resources, we mark the receipt status as "not required", so setting an expected receipt date field doesn't make sense for this mapping. However, it is marked as a eContent required field.
- "manual_po": field doesn't exist in FOLIO
- "qty_ord_elecrtonic": field doesn't exist in FOLIO. Using QUANTITY for now.